requirements.txt
accelerate==1.12.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
av==16.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
filelock==3.20.0
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
peft==0.18.1
pillow==12.0.0
protobuf==6.33.3
psutil==7.2.1
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.15.3
simsimd==6.5.12
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
tensorboardX==2.6.4
tokenizers==0.22.2
torch==2.8.0
torchaudio==2.8.0
torchcodec==0.6.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
